+++ Assets/Mirror/Runtime/NetworkManager.cs
@@ -1,18 +1,13 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using Mirror.Tcp;
 using UnityEngine;
+using UnityEngine.Events;
 using UnityEngine.Rendering;
 using UnityEngine.SceneManagement;
 using UnityEngine.Serialization;
 
 namespace Mirror
 {
-    /// <summary>
-    /// Enumeration of methods of where to spawn player objects in multiplayer games.
-    /// </summary>
-    public enum PlayerSpawnMethod { Random, RoundRobin }
 
     /// <summary>
     /// Enumeration of methods of current Network Manager state at runtime.
@@ -64,26 +59,6 @@ namespace Mirror
         protected Transport transport;
 
         /// <summary>
-        /// A flag to control whether or not player objects are automatically created on connect, and on scene change.
-        /// </summary>
-        [FormerlySerializedAs("m_AutoCreatePlayer")]
-        [Tooltip("Should Mirror automatically spawn the player after scene change?")]
-        public bool autoCreatePlayer = true;
-
-        /// <summary>
-        /// The current method of spawning players used by the NetworkManager.
-        /// </summary>
-        [FormerlySerializedAs("m_PlayerSpawnMethod")]
-        [Tooltip("Round Robin or Random order of Start Position selection")]
-        public PlayerSpawnMethod playerSpawnMethod;
-
-        /// <summary>
-        /// Number of active player objects across all connections on the server.
-        /// <para>This is only valid on the host / server.</para>
-        /// </summary>
-        public int numPlayers => server.connections.Count(kv => kv.Value.identity != null);
-
-        /// <summary>
         /// True if the server or client is started and running
         /// <para>This is set True in StartServer / StartClient, and set False in StopServer / StopClient</para>
         /// </summary>
@@ -109,6 +84,17 @@ namespace Mirror
         //    in other words, we need this to know which mode we are running in
         //    during FinishLoadScene.
         public NetworkManagerMode mode { get; private set; }
+
+        /// <summary>
+        /// This is invoked when a host is started.
+        /// <para>StartHost has multiple signatures, but they all cause this hook to be called.</para>
+        /// </summary>
+        public UnityEvent OnStartHost = new UnityEvent();
+
+        /// <summary>
+        /// This is called when a host is stopped.
+        /// </summary>
+        public UnityEvent OnStopHost = new UnityEvent();
 
         #region Unity Callbacks
 
@@ -151,12 +137,6 @@ namespace Mirror
                 UnityEditor.Undo.RecordObject(gameObject, "Added NetworkClient");
 #endif
             }
-
-            if (server != null && server.playerPrefab != null && server.playerPrefab.GetComponent<NetworkIdentity>() == null)
-            {
-                Debug.LogError("NetworkManager - playerPrefab must have a NetworkIdentity.");
-                server.playerPrefab = null;
-            }
         }
 
         /// <summary>
@@ -169,8 +149,6 @@ namespace Mirror
             // Set the networkSceneName to prevent a scene reload
             // if client connection to server fails.
             networkSceneName = null;
-
-            client.Authenticated.AddListener(OnAuthenticated);
 
             Initialize();
 
@@ -326,7 +304,7 @@ namespace Mirror
             // call OnStartHost AFTER SetupServer. this way we can use
             // NetworkServer.Spawn etc. in there too. just like OnStartServer
             // is called after the server is actually properly started.
-            OnStartHost();
+            OnStartHost.Invoke();
 
             FinishStartHost();
         }
@@ -394,16 +372,7 @@ namespace Mirror
         /// </summary>
         public void StopHost()
         {
-            OnStopHost();
-
-            // TODO try to move DisconnectLocalServer into StopClient(), and
-            // then call StopClient() before StopServer(). needs testing!.
-
-            // DisconnectLocalServer needs to be called so that the host client
-            // receives a DisconnectMessage too.
-            // fixes: https://github.com/vis2k/Mirror/issues/1515
-            NetworkClient.DisconnectLocalServer();
-
+            OnStopHost.Invoke();
             StopClient();
             StopServer();
         }
@@ -423,8 +392,6 @@ namespace Mirror
             // set offline mode BEFORE changing scene so that FinishStartScene
             // doesn't think we need initialize anything.
             mode = NetworkManagerMode.Offline;
-
-            startPositionIndex = 0;
         }
 
         /// <summary>
@@ -510,10 +477,6 @@ namespace Mirror
 
             Transport.activeTransport = transport;
 
-            if (server.playerPrefab != null)
-            {
-                client.RegisterPrefab(server.playerPrefab);
-            }
             // subscribe to the server
             if (server != null)
                 server.Authenticated.AddListener(OnServerAuthenticated);
@@ -575,15 +538,7 @@ namespace Mirror
 
             // notify all clients about the new scene
             server.SendToAll(new SceneMessage { sceneName = newSceneName });
-
-            startPositionIndex = 0;
-            startPositions.Clear();
         }
-
-        // This is only set in ClientChangeScene below...never on server.
-        // We need to check this in OnClientSceneChanged called from FinishLoadSceneClientOnly
-        // to prevent AddPlayer message after loading/unloading additive scenes
-        SceneOperation clientSceneOperation = SceneOperation.Normal;
 
         internal void ClientChangeScene(string newSceneName, SceneOperation sceneOperation = SceneOperation.Normal, bool customHandling = false)
         {
@@ -609,9 +564,6 @@ namespace Mirror
                 FinishLoadScene();
                 return;
             }
-
-            // cache sceneOperation so we know what was done in OnClientSceneChanged called from FinishLoadSceneClientOnly
-            clientSceneOperation = sceneOperation;
 
             switch (sceneOperation)
             {
@@ -766,51 +718,11 @@ namespace Mirror
 
         #endregion
 
-        #region Start Positions
-
-        public int startPositionIndex;
-
-        /// <summary>
-        /// List of transforms populted by NetworkStartPosition components found in the scene.
-        /// </summary>
-        public List<Transform> startPositions = new List<Transform>();
-
-        /// <summary>
-        /// Registers the transform of a game object as a player spawn location.
-        /// <para>This is done automatically by NetworkStartPosition components, but can be done manually from user script code.</para>
-        /// </summary>
-        /// <param name="start">Transform to register.</param>
-        public void RegisterStartPosition(Transform start)
-        {
-            if (LogFilter.Debug) Debug.Log("RegisterStartPosition: (" + start.gameObject.name + ") " + start.position);
-            startPositions.Add(start);
-
-            // reorder the list so that round-robin spawning uses the start positions
-            // in hierarchy order.  This assumes all objects with NetworkStartPosition
-            // component are siblings, either in the scene root or together as children
-            // under a single parent in the scene.
-            startPositions = startPositions.OrderBy(transform => transform.GetSiblingIndex()).ToList();
-        }
-
-        /// <summary>
-        /// Unregisters the transform of a game object as a player spawn location.
-        /// <para>This is done automatically by the <see cref="NetworkStartPosition">NetworkStartPosition</see> component, but can be done manually from user code.</para>
-        /// </summary>
-        /// <param name="start">Transform to unregister.</param>
-        public void UnRegisterStartPosition(Transform start)
-        {
-            if (LogFilter.Debug) Debug.Log("UnRegisterStartPosition: (" + start.gameObject.name + ") " + start.position);
-            startPositions.Remove(start);
-        }
-
-        #endregion
-
         #region Server Internal Message Handlers
 
         void RegisterServerMessages(NetworkConnection connection)
         {
             connection.RegisterHandler<NetworkConnectionToClient, ReadyMessage>(OnServerReadyMessageInternal);
-            connection.RegisterHandler<NetworkConnectionToClient, AddPlayerMessage>(OnServerAddPlayerInternal);
             connection.RegisterHandler<NetworkConnectionToClient, RemovePlayerMessage>(OnServerRemovePlayerMessageInternal);
         }
 
@@ -836,31 +748,6 @@ namespace Mirror
         {
             if (LogFilter.Debug) Debug.Log("NetworkManager.OnServerReadyMessageInternal");
             OnServerReady(conn);
-        }
-
-        void OnServerAddPlayerInternal(NetworkConnection conn, AddPlayerMessage msg)
-        {
-            if (LogFilter.Debug) Debug.Log("NetworkManager.OnServerAddPlayer");
-
-            if (autoCreatePlayer && server.playerPrefab == null)
-            {
-                Debug.LogError("The PlayerPrefab is empty on the NetworkManager. Please setup a PlayerPrefab object.");
-                return;
-            }
-
-            if (autoCreatePlayer && server.playerPrefab.GetComponent<NetworkIdentity>() == null)
-            {
-                Debug.LogError("The PlayerPrefab does not have a NetworkIdentity. Please add a NetworkIdentity to the player prefab.");
-                return;
-            }
-
-            if (conn.identity != null)
-            {
-                Debug.LogError("There is already a player for this connection.");
-                return;
-            }
-
-            OnServerAddPlayer(conn);
         }
 
         void OnServerRemovePlayerMessageInternal(NetworkConnection conn, RemovePlayerMessage msg)
@@ -943,45 +830,6 @@ namespace Mirror
             server.SetClientReady(conn);
         }
 
-        /// <summary>
-        /// Called on the server when a client adds a new player with ClientScene.AddPlayer.
-        /// <para>The default implementation for this function creates a new player object from the playerPrefab.</para>
-        /// </summary>
-        /// <param name="conn">Connection from client.</param>
-        public virtual void OnServerAddPlayer(NetworkConnection conn)
-        {
-            Transform startPos = GetStartPosition();
-            GameObject player = startPos != null
-                ? Instantiate(server.playerPrefab, startPos.position, startPos.rotation)
-                : Instantiate(server.playerPrefab);
-
-            server.AddPlayerForConnection(conn, player);
-        }
-
-        /// <summary>
-        /// This finds a spawn position based on NetworkStartPosition objects in the scene.
-        /// <para>This is used by the default implementation of OnServerAddPlayer.</para>
-        /// </summary>
-        /// <returns>Returns the transform to spawn a player at, or null.</returns>
-        public Transform GetStartPosition()
-        {
-            // first remove any dead transforms
-            startPositions.RemoveAll(t => t == null);
-
-            if (startPositions.Count == 0)
-                return null;
-
-            if (playerSpawnMethod == PlayerSpawnMethod.Random)
-            {
-                return startPositions[UnityEngine.Random.Range(0, startPositions.Count)];
-            }
-            else
-            {
-                Transform startPosition = startPositions[startPositionIndex];
-                startPositionIndex = (startPositionIndex + 1) % startPositions.Count;
-                return startPosition;
-            }
-        }
 
         /// <summary>
         /// Called on the server when a client removes a player.
@@ -1022,27 +870,6 @@ namespace Mirror
         #region Client System Callbacks
 
         /// <summary>
-        /// Called on the client when connected to a server.
-        /// <para>The default implementation of this function sets the client as ready and adds a player. Override the function to dictate what happens when the client connects.</para>
-        /// </summary>
-        /// <param name="conn">Connection to the server.</param>
-        public void OnAuthenticated(NetworkConnectionToServer conn)
-        {
-            // OnClientConnect by default calls AddPlayer but it should not do
-            // that when we have online/offline scenes. so we need the
-            // clientLoadedScene flag to prevent it.
-            if (!clientLoadedScene)
-            {
-                // Ready/AddPlayer is usually triggered by a scene load completing. if no scene was loaded, then Ready/AddPlayer it here instead.
-                if (!client.ready) client.Ready(conn);
-                if (autoCreatePlayer)
-                {
-                    client.AddPlayer();
-                }
-            }
-        }
-
-        /// <summary>
         /// Called on clients when a network error occurs.
         /// </summary>
         /// <param name="conn">Connection to a server.</param>
@@ -1073,14 +900,8 @@ namespace Mirror
         public virtual void OnClientSceneChanged(NetworkConnectionToServer conn)
         {
             // always become ready.
-            if (!client.ready) client.Ready(conn);
-
-            // Only call AddPlayer for normal scene changes, not additive load/unload
-            if (clientSceneOperation == SceneOperation.Normal && autoCreatePlayer && client.localPlayer == null)
-            {
-                // add player if existing one is null
-                client.AddPlayer();
-            }
+            if (!client.ready)
+                client.Ready(conn);
         }
 
         #endregion
@@ -1092,21 +913,10 @@ namespace Mirror
         // from all versions, so users only need to implement this one case.
 
         /// <summary>
-        /// This is invoked when a host is started.
-        /// <para>StartHost has multiple signatures, but they all cause this hook to be called.</para>
-        /// </summary>
-        public virtual void OnStartHost() { }
-
-        /// <summary>
         /// This is invoked when a server is started - including when a host is started.
         /// <para>StartServer has multiple signatures, but they all cause this hook to be called.</para>
         /// </summary>
         public virtual void OnStartServer() { }
-
-        /// <summary>
-        /// This is called when a host is stopped.
-        /// </summary>
-        public virtual void OnStopHost() { }
 
         #endregion
     }