+++ Assets/Mirror/Tests/Runtime/HostTests.cs
@@ -31,7 +31,6 @@ namespace Mirror.Tests
             server = manager.server;
             client = manager.client;
             manager.startOnHeadless = false;
-            manager.autoCreatePlayer = false;
 
             // wait for manager to Start()
             yield return null;